refactor: moves `constants.ts` file to existent `/config/constants.ts`
