+++ actions/auth/logout.ts
@@ -1,12 +1,12 @@
 'use server';
 
-import { constants } from '@/config/constants';
+import { projectConstants } from '@/config/constants';
 import { cookies } from 'next/headers';
 import { redirect } from 'next/navigation';
 
 export async function logout() {
   const cookiesStorage = await cookies();
-  cookiesStorage.delete(constants.accessToken);
+  cookiesStorage.delete(projectConstants.accessToken);
 
   return redirect('/');
 }
