+++ middleware.ts
@@ -1,9 +1,9 @@
 import { cookies } from 'next/headers';
 import type { NextRequest } from 'next/server';
 import { NextResponse } from 'next/server';
-import { constants } from './config/constants';
 
 import { jwtVerify } from 'jose';
+import { projectConstants } from './config/constants';
 
 const publicPaths = ['/', '/auth/sign-in', '/auth/sign-up'];
 
@@ -15,11 +15,11 @@ export async function middleware(request: NextRequest) {
     request.nextUrl.searchParams.get('action') === 'logout';
 
   if (isLogoutAction) {
-    cookieStore.delete(constants.accessToken);
+    cookieStore.delete(projectConstants.accessToken);
     return NextResponse.redirect(new URL('/', request.url));
   }
 
-  const token = cookieStore.get(constants.accessToken)?.value;
+  const token = cookieStore.get(projectConstants.accessToken)?.value;
 
   if (token && publicPaths.includes(pathname)) {
     return NextResponse.redirect(new URL('/dashboard', request.url));
@@ -35,7 +35,7 @@ export async function middleware(request: NextRequest) {
     try {
       await jwtVerify(token, new TextEncoder().encode(process.env.JWT_SECRET!));
     } catch {
-      cookieStore.delete(constants.accessToken);
+      cookieStore.delete(projectConstants.accessToken);
       return NextResponse.redirect(new URL('/', request.url));
     }
   }
